requirements.txt
pandas
scikit-learn
fastapi
uvicorn
requests
joblib
matplotlib
seaborn
jupyter
azure-storage-blob